Moved PipelineStage to dedicated source file
